requirements.txt
# Python dependencies for the IDS project

watchdog

python-nmap

psutil